Optimize proximity checking algorithm and reducing proximity scores

+++ src/RelevanceFunction.h
@@ -9,7 +9,7 @@
 float getWordRelevanceScore(int wordCount, int totalWords, int filesContainingWord);
 
 // Calculates bonus marks based on the proximity of word positions in the file
-int calculateBonusMarks(int pos1, int pos2);
+float calculateBonusMarks(int pos1, int pos2);
 
 // Returns the search relevance score for the file based on multiple words searched and their positions
 float getSearchRelevanceScore(BookInfo bookInfos);
